requirements.txt
streamlit
google-generativeai
firebase-admin
python-dotenv
openai-whisper
